Move Conversion feature to vertical-slice folders
Co-locates the ConverterController, converter DTOs (UpdateVideoInfoRequest,
VideoToTransformResponse, GetPublishSasResponse, UploadConvertedVideoResponse),
and the per-feature ConversionRoutes constants under Features/Conversion/.

The converter-facing methods on VideoUploaderService stay in
Application/Features/Videos/ because the same class also serves the user-
facing upload flow; splitting it would exceed the folder-reorg scope.

Removes ApiEndpoints.Converter (constants now in ConversionRoutes). Empties
the Contracts/Requests/ and Contracts/Responses/ folders — all DTOs now live
under Features/<Name>/. Cleans up dangling using directives in API mappers,
mobile library, MAUI app, and tests that referenced the old namespaces.
