+++ src/backend/TB.DanceDance.API.Contracts/Features/Conversion/UploadConvertedVideoResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace TB.DanceDance.API.Contracts.Responses
+namespace TB.DanceDance.API.Contracts.Features.Conversion
 {
     public class UploadConvertedVideoResponse
     {
