+++ src/backend/TB.DanceDance.API/Mappers/ContractMappers.cs
@@ -2,8 +2,6 @@
 using TB.DanceDance.API.Contracts.Features.Events;
 using TB.DanceDance.API.Contracts.Features.Videos;
 using TB.DanceDance.API.Contracts.Models;
-using TB.DanceDance.API.Contracts.Requests;
-using TB.DanceDance.API.Contracts.Responses;
 using TB.DanceDance.API.Extensions;
 
 namespace TB.DanceDance.API.Mappers;
