+++ src/backend/TB.DanceDance.Services.Converter.Deamon/IDanceDanceApiClient.cs
@@ -1,5 +1,4 @@
-﻿using TB.DanceDance.API.Contracts.Requests;
-using TB.DanceDance.API.Contracts.Responses;
+﻿using TB.DanceDance.API.Contracts.Features.Conversion;
 
 namespace TB.DanceDance.Services.Converter.Deamon;
