+++ src/mobile/TB.DanceDance.Mobile.Library/Services/DanceApi/VideoUploader.cs
@@ -2,7 +2,6 @@
 using System.Threading.Channels;
 using TB.DanceDance.API.Contracts.Features.Videos;
 using TB.DanceDance.API.Contracts.Models;
-using TB.DanceDance.API.Contracts.Requests;
 using TB.DanceDance.Mobile.Library.Data;
 using TB.DanceDance.Mobile.Library.Data.Models.Storage;
 using TB.DanceDance.Mobile.Library.Services.Network;
