+++ src/tests/TB.DanceDance.Mobile.Tests/IntegrationTests/VideoUploaderTests.cs
@@ -4,7 +4,6 @@ using NSubstitute;
 using System.Threading.Channels;
 using TB.DanceDance.API.Contracts.Features.Videos;
 using TB.DanceDance.API.Contracts.Models;
-using TB.DanceDance.API.Contracts.Requests;
 using TB.DanceDance.Mobile.Library.Data;
 using TB.DanceDance.Mobile.Library.Data.Models.Storage;
 using TB.DanceDance.Mobile.Library.Services.DanceApi;
